requirements.txt
streamlit
numpy
opencv-python-headless
scikit-learn
pillow
matplotlib
